02 Web Application with Spring Boot: 036 Step 25 Exception Handling

+++ spring-boot-first-web-application/src/main/java/com/gekif/springboot/web/springbootfirstwebapplication/controller/ToDoController.java
@@ -64,6 +64,8 @@ public class ToDoController {
 
     @RequestMapping(value= "/delete-todo", method = RequestMethod.GET)
     public String deleteTodo(@RequestParam int id) {
+        if (id == 1)
+            throw new RuntimeException("Something went wrong");
         service.deleteTodo(id);
         return "redirect:/list-todos";
     }
